app/src/main/java/com/example/user/week07_day01_movie_rankings/Movie.java
package com.example.user.week07_day01_movie_rankings;

/**
 * Created by user on 03/07/2017.
 */

public class Movie {

    private String title;
    private String genre;
    private int ranking;

    public Movie(String title, String genre, int ranking) {
        this.title = title;
        this.genre = genre;
        this.ranking = ranking;
    }

    public String getTitle() {
        return this.title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getGenre() {
        return this.genre;
    }

    public void setGenre(String genre) {
        this.genre = genre;
    }

    public int getRanking() {
        return ranking;
    }

    public void setRanking(int ranking) {
        this.ranking = ranking;
    }

    public String toString() {
        return "Title: '" + getTitle() + "', Genre: '" + getGenre() + "', Ranking: " + getRanking();
    }

}
